Implement basic routing structure
Add pages for listing and viewing specific communties and campaigns. Use
a mock database with fake data.

These initial components serve as a base to build on top of but
certainly don't do everything right. For example, we'll likely want to
replace the direct DB calls within the components with calls to an
abstraction layer that returns all of the data a specific page needs.