Add basic annotation features to map component
Enable passing annotations to the map, viewing their details, deleting
them, and adding new ones. Use Mapbox's native popup functionality to
show annotations in the map. Focus on functionality, not styling, for
now.

Several needed changes to the current select mode are not yet
implemented:

- Disable selecting multiple annotations at once (or handle it well).
- Disable dragging annotations.
- Disable clicking on an annotation a second time to edit it's vertices.

We'll certainly want some of that functionality eventually, but how it
should be implemented remains undecided. For now, implement a base set
of functionality that we can build upon.

Note that some of the above behaviors currently cause crashes.