Style site header
Use placeholder text for the language and account menus for now.

Also disable the react/require-default-props eslint rule because I don't
like it. TypeScript already requires us to handle the case where a prop
is missing and therefore undefined, so I don't see the need to
explicitly set a default as well. Undefined is the default.

Closes #5.