Setup Playwright for testing
Go for an end-to-end testing framework because I've found unit testing
React components to be fragile and not particularly useful. We'll
probably add a unit testing framework for non-React code once we have
more of it.

Forego setting up the webkit browser for now because it's a pain on my
computer. If we start getting webkit-specific bugs, we'll re-evaluate.