Make header render more self-sufficiently
Include the header in the root layout so it gets added to all pages.
Parse the pathname to determine whether the breadcrumb should be
displayed and to get the data it needs. (Note that the pathname must be
obtained with usePathname() rather than relying on route params because
the root layout doesn't get any route params; see
https://github.com/vercel/next.js/discussions/49507.)

Leave a few TODOs about DB access in the breadcrumb that will be
resolved when switching to a Supabase-backed database.