Implement community task page
A task list is shown next to the map. Tasks can be filterd by priority,
status, or campaign. They can be sorted by the same fields, plus the
date, in ascending or descending order. They can also be "searched",
effectively filtering them based on a case-insensitive string match
within the name or description.