Fix creating annotations with campaign filtering
When a new annotation is created, it isn't initially associated with any
campaign. As a result, it is immediately removed from the map (and its
popup is closed) by the filtering.

Fix this by adding a "No campaign" entry to the map filter menu for
annotations that are not associated with any campaign. Also ensure that
the annotation popup doesn't get closed even if the provided annotations
no longer include the selected one, to handle cases where a new
annotation is selected but the "No campaign" filter is disabled.