Verify authorization
Only allow users to interact with data that that's tied to a community
that they've either created or are a collaborator on. Seed the database
with several test users to improve development.

Since this change makes most pages useless without a user account,
redirect to the login page from any page when not logged in.

Closes: #26