Store Mapbox map in state
I've been running into issues with the useDraw hook trying to make
changes after the map has already been removed. To fix this issue, give
the various map hooks a reference to the map object itself, not just a
React ref. Store the map object in state so that changing it triggers a
rerender.