Add version route
It's nice to be able to easily tell which version is running in prod.
Ideally we would pass the version in as a build arg, but I can't figure
out how to do that in OpenShift so we're just going to include the .git
directory in the container as well. Whatever.