Ensure that filter always has all enabled values
Since the enabled values are stored in state and updated with a
useEffect hook, they can get behind the other useFilter return values,
causing the filter menu to throw an error because it doesn't know the
enabled state for some entries. Address this issue by always returning a
full enabled state, using the same defaults that will be applied when
the useEffect hook updates the actual state.

Fixes: #43