Improve date input behavior
The date input is weird -- it triggers the change handler whenever
anything is typed in it, but if what is typed causes the date to become
partial (i.e., because a leading zero is typed in one of the
sub-fields), then the value is null in the handler. If we make the input
controlled, this null value will get reflected back and clear the entire
date input. Change the input's value when it loses focus instead to
prevent these updates when the user is halfway through entering the
date.

Fixes: #44