Don't delete annotations in progress from map
When drawing a new annotation on the map, most of the draw modes begin
by adding a feature to the map and then updating it throughout the
drawing process. Whenever the list of annotations changes, though, we
reset the features to those derived from the annotations. If the list
changes while drawing is in progress, the in progress feature
disappears. Prevent this issue by keeping any annotations that weren't
added as a result of the annotations.

Fixes: #48