Add row-level security to profiles
I got an email from Supabase reminding me that RLS was off for the
profiles table. Add some basic policies. We need to allow all
authenticated users to view all profiles so that they can read the admin
profile when creating a new community. (I suppose we could tighten the
permissions by storing the admin user id in the DB somewhere and then
only allowing access to that one, but it feels unnecessary at the
moment.)