Reset track or campaign id after save and delete
We were correctly reseting the id after closing the campaign or task dialog,
but not after a save or a delete action. This slightly affected the tracking
behavior when we would add two new tasks or campaigns in a row because the
id of the first one would persist in the new one.