Handle missing user profiles gracefully
When creating a new community, we fetch the user profiles of the current
user and the admin user in order to display their names. Gracefully
handle the case where one of these users doesn't have a profile by using
a default name. (A user can only be missing a profile if the user was
added to Supabase manually without a corresponding profile -- this issue
shouldn't occur in prod.)

Fixes: #73